[level Lv.1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.17 ms, Memory: 31.1 MB -BaekjoonHub

+++ 프로그래머스/Lv.1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.cs
@@ -1,7 +1,7 @@
 public class Solution {
     public long[] solution(int x, int n) {
         long[] answer = new long[n];
-        long tmp = 0;
+        int tmp = 0;
         if (x >= -10000000  || x <= 10000000  || n < 1000)
         {
